move pthread_queue.h to include, fixing build bustage on Mac OS X
